--- jetbrains-ultimate/src/software/aws/toolkits/jetbrains/services/lambda/nodejs/NodeJsRuntimeDebugSupport.kt
@@ -1,18 +1,0 @@
-// Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package software.aws.toolkits.jetbrains.services.lambda.nodejs
-
-import com.intellij.execution.runners.ExecutionEnvironment
-import com.intellij.xdebugger.XDebugProcessStarter
-import software.aws.toolkits.jetbrains.services.lambda.execution.sam.RuntimeDebugSupport
-import software.aws.toolkits.jetbrains.services.lambda.execution.sam.SamRunningState
-
-class NodeJsRuntimeDebugSupport : RuntimeDebugSupport {
-    override suspend fun createDebugProcess(
-        environment: ExecutionEnvironment,
-        state: SamRunningState,
-        debugHost: String,
-        debugPorts: List<Int>
-    ): XDebugProcessStarter = NodeJsDebugUtils.createDebugProcess(state, debugHost, debugPorts)
-}
